requirements.txt
llama-cpp-python==0.2.6
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
pydantic==2.6.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0